# server/api/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Visora AI P1", "version": "0.2.0", "description": "本地开发账号服务与运营后台。写请求必须通过浏览器 Origin 精确白名单验证，缺失或不匹配返回403 FORBIDDEN。写方法包含 POST/PATCH/PUT/DELETE，四者都要求 Origin、X-Visora-Request 与 Content-Type 三道校验。浏览器只调用本 origin 的 /api；开发用户端/Admin统一使用127.0.0.1，不混用localhost。代理保留 Origin、Cookie 和 Set-Cookie，不让浏览器直接请求API_TARGET。项目云存储、平台生成任务与积分消耗结算仍未接入；积分流水中的 consume 类型由后续生成迭代写入，当前没有写入方。" },
  "security": [{ "session": [] }],
  "paths": {
...
    "/api/admin/models": {
      "get": {
        "operationId": "listModels", "description": "仅管理员；平台模型目录，createdAt/_id 倒序。目录只放报价与启停，不含供应商密钥。",
        "parameters": [{ "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "$ref": "#/components/responses/ModelPage" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "operationId": "createModel", "description": "仅管理员；名称唯一，重复返回 409 NAME_EXISTS。kind 仅接受 image/video/text/audio。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ModelCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Model" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/models/{id}": {
      "patch": {
        "operationId": "updateModel", "description": "仅管理员；局部更新，至少要带一个字段，否则 400。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ModelUpdate" },
        "responses": { "200": { "$ref": "#/components/responses/Model" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "headers": { "NoStore": { "schema": { "type": "string", "const": "no-store" } } },
    "securitySchemes": { "session": { "type": "apiKey", "in": "cookie", "name": "visora_session" } },
    "parameters": {
      "RequestHeader": { "name": "X-Visora-Request", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string" } },
      "PathID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
    },
    "requestBodies": {
...
      "ModelCreate": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "kind", "pointsCost"], "additionalProperties": false, "properties": { "name": { "type": "string", "maxLength": 60 }, "kind": { "type": "string", "enum": ["image", "video", "text", "audio"] }, "pointsCost": { "type": "integer", "minimum": 0 } } } } } },
      "ModelUpdate": { "required": true, "content": { "application/json": { "schema": { "type": "object", "minProperties": 1, "additionalProperties": false, "properties": { "name": { "type": "string", "maxLength": 60 }, "kind": { "type": "string", "enum": ["image", "video", "text", "audio"] }, "pointsCost": { "type": "integer", "minimum": 0 }, "status": { "type": "string", "enum": ["active", "disabled"] } } } } } },
...
    },
    "schemas": {
...
      "AIModel": { "type": "object", "required": ["id", "name", "kind", "pointsCost", "status", "calls", "createdAt", "updatedAt"], "additionalProperties": false, "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "kind": { "type": "string", "enum": ["image", "video", "text", "audio"] }, "pointsCost": { "type": "integer", "description": "每次调用的积分报价。" }, "status": { "type": "string", "enum": ["active", "disabled"] }, "calls": { "type": "integer", "description": "累计调用次数。平台生成尚未接入，当前恒为 0。" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } } },
...
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string", "description": "400 INVALID_INPUT；401 INVALID_CREDENTIALS（登录）/UNAUTHENTICATED（会话）；403 FORBIDDEN（角色/Origin/header/Content-Type）；404 NOT_FOUND/MEMBER_NOT_FOUND/PLAN_NOT_FOUND；405 METHOD_NOT_ALLOWED；409 EMAIL_EXISTS/NAME_EXISTS/USER_PROTECTED/TEAM_INACTIVE/MEMBER_EXISTS/MEMBER_HAS_POINTS/INSUFFICIENT_POINTS/ALREADY_DECIDED/SELF_ACTION/LAST_ADMIN；413 PAYLOAD_TOO_LARGE；503 SERVICE_UNAVAILABLE。", "examples": ["INVALID_INPUT", "INVALID_CREDENTIALS", "UNAUTHENTICATED", "FORBIDDEN", "NOT_FOUND", "MEMBER_NOT_FOUND", "PLAN_NOT_FOUND", "METHOD_NOT_ALLOWED", "EMAIL_EXISTS", "NAME_EXISTS", "USER_PROTECTED", "TEAM_INACTIVE", "MEMBER_EXISTS", "MEMBER_HAS_POINTS", "INSUFFICIENT_POINTS", "ALREADY_DECIDED", "SELF_ACTION", "LAST_ADMIN", "PAYLOAD_TOO_LARGE", "SERVICE_UNAVAILABLE"] }, "message": { "type": "string" } } } } }
    },
    "responses": {
...
      "Model": { "description": "模型", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["model"], "properties": { "model": { "$ref": "#/components/schemas/AIModel" } } } } } },
...
      "ModelPage": { "description": "模型分页", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/AIModel" } }, "nextCursor": { "type": ["string", "null"] } } } } } },
...
      "Error": { "description": "受控错误；不回显数据库错误或凭据", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}